* Headers comparison should be case insensitive * Moving MessageSystemAttribute to rest as it is used only in this module
